Optional append domain name to username when authenticate
When having special characters like '@'  in the username, the auto creation of user fails due to username restrictions. With the configuration option '$wgImapAuthorizationAppendDomain' you are able to optionally append a domain name to the username. So the username in wiki might be "f.surename" and the IMAP username "f.surename@domain.name".